+++ src/minecraft/ip/industrialProcessing/items/guide/gui/machines/components/GuideMachineFurnaceRecipeDetails.java
@@ -8,9 +8,6 @@ import net.minecraft.client.gui.Gui;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.IRecipe;
-import net.minecraft.item.crafting.ShapedRecipes;
-import net.minecraft.item.crafting.ShapelessRecipes;
 import net.minecraft.util.ResourceLocation;
 
 public class GuideMachineFurnaceRecipeDetails extends GuideMachineRecipeDetails {
@@ -57,10 +54,10 @@ public class GuideMachineFurnaceRecipeDetails extends GuideMachineRecipeDetails 
 
 		Tessellator tessellator = Tessellator.instance;
 		tessellator.startDrawingQuads();
-		tessellator.addVertexWithUV((double) (x + 0), (double) (y + h), (double) this.zLevel, 0, 1);
-		tessellator.addVertexWithUV((double) (x + w), (double) (y + h), (double) this.zLevel, 1, 1);
-		tessellator.addVertexWithUV((double) (x + w), (double) (y + 0), (double) this.zLevel, 1, 0);
-		tessellator.addVertexWithUV((double) (x + 0), (double) (y + 0), (double) this.zLevel, 0, 0);
+		tessellator.addVertexWithUV(x + 0, y + h, this.zLevel, 0, 1);
+		tessellator.addVertexWithUV(x + w, y + h, this.zLevel, 1, 1);
+		tessellator.addVertexWithUV(x + w, y + 0, this.zLevel, 1, 0);
+		tessellator.addVertexWithUV(x + 0, y + 0, this.zLevel, 0, 0);
 		tessellator.draw();
 	}
