+++ src/minecraft/ip/industrialProcessing/power/plants/ModelBoiler.java
@@ -1,6 +1,5 @@
 package ip.industrialProcessing.power.plants;
 
-import ip.industrialProcessing.client.render.ConnectionState;
 import ip.industrialProcessing.client.render.ModelAnimatedFluidMachine;
 import net.minecraft.client.model.ModelRenderer;
 import net.minecraft.client.renderer.Tessellator;
@@ -91,20 +90,20 @@ public class ModelBoiler extends ModelAnimatedFluidMachine {
         tessellator.startDrawingQuads();
         tessellator.setBrightness(lightLevel);
         tessellator.setNormal(0, -1, 0);
-        tessellator.addVertexWithUV((double) (x + 0) * f, (double) y * f, (double) (z + h1) * f, (double) u0, (double) v1);
-        tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + 0) * f, (double) u1, (double) v0);
-        tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + h) * f, (double) u1, (double) v3);
-        tessellator.addVertexWithUV((double) (x + 0) * f, (double) y * f, (double) (z + h2) * f, (double) u0, (double) v2);
+        tessellator.addVertexWithUV((double) (x + 0) * f, (double) y * f, (double) (z + h1) * f, u0, v1);
+        tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + 0) * f, u1, v0);
+        tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + h) * f, u1, v3);
+        tessellator.addVertexWithUV((double) (x + 0) * f, (double) y * f, (double) (z + h2) * f, u0, v2);
 
-        tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + 0) * f, (double) u1, (double) icon.getMinV());
-        tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + 0) * f, (double) u2, (double) icon.getMinV());
-        tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + h) * f, (double) u2, (double) icon.getMaxV());
-        tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + h) * f, (double) u1, (double) icon.getMaxV());
+        tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + 0) * f, u1, icon.getMinV());
+        tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + 0) * f, u2, icon.getMinV());
+        tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + h) * f, u2, icon.getMaxV());
+        tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + h) * f, u1, icon.getMaxV());
 
-        tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + 0) * f, (double) u2, (double) v0);
-        tessellator.addVertexWithUV((double) (x + w) * f, (double) y * f, (double) (z + h1) * f, (double) u3, (double) v1);
-        tessellator.addVertexWithUV((double) (x + w) * f, (double) y * f, (double) (z + h2) * f, (double) u3, (double) v2);
-        tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + h) * f, (double) u2, (double) v3);
+        tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + 0) * f, u2, v0);
+        tessellator.addVertexWithUV((double) (x + w) * f, (double) y * f, (double) (z + h1) * f, u3, v1);
+        tessellator.addVertexWithUV((double) (x + w) * f, (double) y * f, (double) (z + h2) * f, u3, v2);
+        tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + h) * f, u2, v3);
         tessellator.draw();
 
         float vOffset = (icon.getMaxV() - icon.getMinV()) * (1 - fluidPercentage);
@@ -114,15 +113,15 @@ public class ModelBoiler extends ModelAnimatedFluidMachine {
         tessellator.setBrightness(lightLevel);
         tessellator.addVertexWithUV((double) (x + 4) * f, (double) (y) * f, (double) (z + h) * f, (double) icon.getMinU() + uOffset, (double) icon.getMinV() + vOffset);
         tessellator.addVertexWithUV((double) (x + w - 4) * f, (double) (y) * f, (double) (z + h) * f, (double) icon.getMaxU() - uOffset, (double) icon.getMinV() + vOffset);
-        tessellator.addVertexWithUV((double) (x + w - 4) * f, (double) (yBot) * f, (double) (z + h) * f, (double) icon.getMaxU() - uOffset, (double) icon.getMaxV());
-        tessellator.addVertexWithUV((double) (x + 4) * f, (double) (yBot) * f, (double) (z + h) * f, (double) icon.getMinU() + uOffset, (double) icon.getMaxV());
+        tessellator.addVertexWithUV((double) (x + w - 4) * f, (double) (yBot) * f, (double) (z + h) * f, (double) icon.getMaxU() - uOffset, icon.getMaxV());
+        tessellator.addVertexWithUV((double) (x + 4) * f, (double) (yBot) * f, (double) (z + h) * f, (double) icon.getMinU() + uOffset, icon.getMaxV());
         
 
         tessellator.setNormal(0, 0, 1);
         tessellator.addVertexWithUV((double) (x + w - 4) * f, (double) (y) * f, (double) (z + 0) * f, (double) icon.getMinU() + uOffset, (double) icon.getMinV() + vOffset);
         tessellator.addVertexWithUV((double) (x +  4) * f, (double) (y) * f, (double) (z + 0) * f, (double) icon.getMaxU() - uOffset, (double) icon.getMinV() + vOffset);
-        tessellator.addVertexWithUV((double) (x +  4) * f, (double) (yBot) * f, (double) (z + 0 ) * f, (double) icon.getMaxU() - uOffset, (double) icon.getMaxV());
-        tessellator.addVertexWithUV((double) (x + w - 4) * f, (double) (yBot) * f, (double) (z + 0 ) * f, (double) icon.getMinU() + uOffset, (double) icon.getMaxV());
+        tessellator.addVertexWithUV((double) (x +  4) * f, (double) (yBot) * f, (double) (z + 0 ) * f, (double) icon.getMaxU() - uOffset, icon.getMaxV());
+        tessellator.addVertexWithUV((double) (x + w - 4) * f, (double) (yBot) * f, (double) (z + 0 ) * f, (double) icon.getMinU() + uOffset, icon.getMaxV());
         tessellator.draw();
 
     }
