+++ src/minecraft/ip/industrialProcessing/power/plants/ModelTurbine.java
@@ -1,9 +1,7 @@
 package ip.industrialProcessing.power.plants;
 
 import ip.industrialProcessing.client.render.ModelAnimatedFluidMachine;
-import ip.industrialProcessing.power.plants.models.GeneratorAxle;
 import ip.industrialProcessing.power.plants.models.TurbineAxle;
-import net.minecraft.client.model.ModelRenderer;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.Icon;
@@ -35,10 +33,10 @@ public class ModelTurbine extends ModelAnimatedFluidMachine {
         tessellator.startDrawingQuads();
         tessellator.setBrightness(lightLevel);
         tessellator.setNormal(0, -1, 0);
-        tessellator.addVertexWithUV((double) (par1 + 0) * f, (double) par3 * f, (double) (par2 + 0) * f, (double) icon.getMinU(), (double) icon.getMinV());
-        tessellator.addVertexWithUV((double) (par1 + par4) * f, (double) par3 * f, (double) (par2 + 0) * f, (double) icon.getMaxU(), (double) icon.getMinV());
-        tessellator.addVertexWithUV((double) (par1 + par4) * f, (double) par3 * f, (double) (par2 + par5) * f, (double) icon.getMaxU(), (double) icon.getMaxV());
-        tessellator.addVertexWithUV((double) (par1 + 0) * f, (double) par3 * f, (double) (par2 + par5) * f, (double) icon.getMinU(), (double) icon.getMaxV());
+        tessellator.addVertexWithUV((double) (par1 + 0) * f, (double) par3 * f, (double) (par2 + 0) * f, icon.getMinU(), icon.getMinV());
+        tessellator.addVertexWithUV((double) (par1 + par4) * f, (double) par3 * f, (double) (par2 + 0) * f, icon.getMaxU(), icon.getMinV());
+        tessellator.addVertexWithUV((double) (par1 + par4) * f, (double) par3 * f, (double) (par2 + par5) * f, icon.getMaxU(), icon.getMaxV());
+        tessellator.addVertexWithUV((double) (par1 + 0) * f, (double) par3 * f, (double) (par2 + par5) * f, icon.getMinU(), icon.getMaxV());
         tessellator.draw();
     }
