+++ src/minecraft/ip/industrialProcessing/transport/fluids/ModelTank.java
@@ -2,15 +2,11 @@ package ip.industrialProcessing.transport.fluids;
 
 import org.lwjgl.opengl.GL11;
 
-import net.minecraft.client.model.ModelRenderer;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.Icon;
-import ip.industrialProcessing.api.rendering.ModelMachine;
 import ip.industrialProcessing.client.render.ConnectionState;
-import ip.industrialProcessing.client.render.ModelConnected;
 import ip.industrialProcessing.client.render.ModelConnectedFluid;
-import ip.industrialProcessing.client.render.ModelStateMachine;
 
 public class ModelTank extends ModelConnectedFluid {
 
@@ -69,20 +65,20 @@ public class ModelTank extends ModelConnectedFluid {
                 tessellator.startDrawingQuads();
                 tessellator.setBrightness(lightLevel);
                 tessellator.setNormal(0, -1, 0);
-                tessellator.addVertexWithUV((double) (x + 0) * f, (double) y * f, (double) (z + h1) * f, (double) u0, (double) v1);
-                tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + 0) * f, (double) u1, (double) v0);
-                tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + h) * f, (double) u1, (double) v3);
-                tessellator.addVertexWithUV((double) (x + 0) * f, (double) y * f, (double) (z + h2) * f, (double) u0, (double) v2);
+                tessellator.addVertexWithUV((double) (x + 0) * f, (double) y * f, (double) (z + h1) * f, u0, v1);
+                tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + 0) * f, u1, v0);
+                tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + h) * f, u1, v3);
+                tessellator.addVertexWithUV((double) (x + 0) * f, (double) y * f, (double) (z + h2) * f, u0, v2);
 
-                tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + 0) * f, (double) u1, (double) icon.getMinV());
-                tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + 0) * f, (double) u2, (double) icon.getMinV());
-                tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + h) * f, (double) u2, (double) icon.getMaxV());
-                tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + h) * f, (double) u1, (double) icon.getMaxV());
+                tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + 0) * f, u1, icon.getMinV());
+                tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + 0) * f, u2, icon.getMinV());
+                tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + h) * f, u2, icon.getMaxV());
+                tessellator.addVertexWithUV((double) (x + w1) * f, (double) y * f, (double) (z + h) * f, u1, icon.getMaxV());
 
-                tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + 0) * f, (double) u2, (double) v0);
-                tessellator.addVertexWithUV((double) (x + w) * f, (double) y * f, (double) (z + h1) * f, (double) u3, (double) v1);
-                tessellator.addVertexWithUV((double) (x + w) * f, (double) y * f, (double) (z + h2) * f, (double) u3, (double) v2);
-                tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + h) * f, (double) u2, (double) v3);
+                tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + 0) * f, u2, v0);
+                tessellator.addVertexWithUV((double) (x + w) * f, (double) y * f, (double) (z + h1) * f, u3, v1);
+                tessellator.addVertexWithUV((double) (x + w) * f, (double) y * f, (double) (z + h2) * f, u3, v2);
+                tessellator.addVertexWithUV((double) (x + w2) * f, (double) y * f, (double) (z + h) * f, u2, v3);
                 tessellator.draw();
 
             }
@@ -93,8 +89,8 @@ public class ModelTank extends ModelConnectedFluid {
             tessellator.setBrightness(lightLevel);
             tessellator.addVertexWithUV((double) (x + 4) * f, (double) (y) * f, (double) (z + h) * f, (double) icon.getMinU() + uOffset, (double) icon.getMinV() + vOffset);
             tessellator.addVertexWithUV((double) (x + w - 4) * f, (double) (y) * f, (double) (z + h) * f, (double) icon.getMaxU() - uOffset, (double) icon.getMinV() + vOffset);
-            tessellator.addVertexWithUV((double) (x + w - 4) * f, (double) (yBot) * f, (double) (z + h) * f, (double) icon.getMaxU() - uOffset, (double) icon.getMaxV());
-            tessellator.addVertexWithUV((double) (x + 4) * f, (double) (yBot) * f, (double) (z + h) * f, (double) icon.getMinU() + uOffset, (double) icon.getMaxV());
+            tessellator.addVertexWithUV((double) (x + w - 4) * f, (double) (yBot) * f, (double) (z + h) * f, (double) icon.getMaxU() - uOffset, icon.getMaxV());
+            tessellator.addVertexWithUV((double) (x + 4) * f, (double) (yBot) * f, (double) (z + h) * f, (double) icon.getMinU() + uOffset, icon.getMaxV());
             tessellator.draw();
             GL11.glDisable(GL11.GL_BLEND);
         }
